use the largest art available in the album dir
